Atual. do pacote DAO e implementação do BD
Atualizei o package DAO. Alterei a classe LeitorDAO, transformando-a em uma interface, com os CRUDs a serem implementados e, junto a ela, foi criada a classe LeitorDAOImplementation, onde ocorre a conexão com o banco de dados, de fato. Também atualizei a classe Conexao, criando a implementação e a conexão com o banco de dados. Por fim, criei o arquivo database.sql, que conteém o scrpit do BD e de suas tabelas e realizei algumas correções na classe Leitor, do package Entity.

+++ src/main/java/edu/curso/entity/Autor.java
@@ -18,7 +18,7 @@ public class Autor {
         this.telefone = telefone;
     }
 
-    public long getCodigo() {
+    public int getCodigo() {
         return codigo;
     }
